// app/Types/types.ts
export type typesOfData = {
  id: string;
  name: string;
  creator: string;
  price: number;
  views: string;
  comments: number;
  image: any;
  date: string;
  topBid: string;
  address: string;
  tokenId: number;
  tokenSt: string;
  blockchain: string;
  avatars: {
    id: string;
    image: any;
  }[];
}[];